moved effect to content-section.js
